# backend/app/api/contracts/step2-unified-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TMDB 海报墙",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/quark/search/tmdb/{tmdb_id}": {
      "get": {
        "tags": [
          "quark"
        ],
        "summary": "通过TMDB ID搜索夸克资源",
        "description": "通过TMDB ID搜索夸克资源",
        "operationId": "search_by_tmdb_id_api_quark_search_tmdb__tmdb_id__get",
        "parameters": [
          {
            "name": "tmdb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmdb Id"
            }
          },
          {
            "name": "media_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "媒体类型，可选值：movie, tv",
              "default": "movie",
              "title": "Media Type"
            },
            "description": "媒体类型，可选值：movie, tv"
          },
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "最大结果数量",
              "default": 20,
              "title": "Max Results"
            },
            "description": "最大结果数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SearchData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quark/search/title": {
      "get": {
        "tags": [
          "quark"
        ],
        "summary": "通过标题搜索夸克资源",
        "description": "通过标题搜索夸克资源",
        "operationId": "search_by_title_api_quark_search_title_get",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "搜索标题",
              "title": "Title"
            },
            "description": "搜索标题"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "年份",
              "title": "Year"
            },
            "description": "年份"
          },
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "最大结果数量",
              "default": 20,
              "title": "Max Results"
            },
            "description": "最大结果数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SearchData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quark/transfer": {
      "post": {
        "tags": [
          "quark"
        ],
        "summary": "保存资源到网盘",
        "description": "保存分享资源到网盘并执行 Emby v1.6 规范重命名。",
        "operationId": "transfer_resource_api_quark_transfer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_QuarkTransferData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponse_QuarkTransferData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/QuarkTransferData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[QuarkTransferData]"
      },
      "ApiResponse_SearchData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/SearchData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[SearchData]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaDto": {
        "properties": {
          "tmdb_id": {
            "type": "integer",
            "title": "Tmdb Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_title": {
            "type": "string",
            "title": "Original Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "overview": {
            "type": "string",
            "title": "Overview"
          },
          "poster_path": {
            "type": "string",
            "title": "Poster Path"
          },
          "backdrop_path": {
            "type": "string",
            "title": "Backdrop Path"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          }
        },
        "type": "object",
        "required": [
          "tmdb_id",
          "title",
          "original_title",
          "year",
          "rating",
          "overview",
          "poster_path",
          "backdrop_path",
          "media_type"
        ],
        "title": "MediaDto",
        "description": "媒体信息DTO"
      },
...
      "QuarkTransferData": {
        "properties": {
          "saved_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Saved Files"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id"
          },
          "collection_created": {
            "type": "boolean",
            "title": "Collection Created",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "saved_files",
          "task_id"
        ],
        "title": "QuarkTransferData"
      },
...
      "ResourceDto": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "link": {
            "type": "string",
            "title": "Link"
          },
          "overall_score": {
            "type": "number",
            "title": "Overall Score"
          },
          "quality_level": {
            "type": "string",
            "title": "Quality Level"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "codec": {
            "type": "string",
            "title": "Codec"
          },
          "is_best": {
            "type": "boolean",
            "title": "Is Best"
          },
          "normalized_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Name"
          },
          "Conf": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conf"
          },
          "Qual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qual"
          },
          "alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alpha"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "size_gb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Gb"
          },
          "C_text": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "C Text"
          },
          "C_intent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "C Intent"
          },
          "C_plaus": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "C Plaus"
          },
          "P": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P"
          },
          "R": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "R"
          }
        },
        "type": "object",
        "required": [
          "name",
          "link",
          "overall_score",
          "quality_level",
          "resolution",
          "codec",
          "is_best"
        ],
        "title": "ResourceDto",
        "description": "资源DTO"
      },
      "SearchData": {
        "properties": {
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/ResourceDto"
            },
            "type": "array",
            "title": "Resources"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "query_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Time"
          }
        },
        "type": "object",
        "required": [
          "resources",
          "total"
        ],
        "title": "SearchData"
      },
...
      "TransferRequest": {
        "properties": {
          "link": {
            "type": "string",
            "title": "Link"
          },
          "to_dir_fid": {
            "type": "string",
            "title": "To Dir Fid",
            "default": "0"
          },
          "to_dir_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Dir Name"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type",
            "default": "movie"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "tmdb_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Id"
          },
          "poster_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Path"
          },
          "backdrop_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Path"
          },
          "resource_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Name"
          }
        },
        "type": "object",
        "required": [
          "link"
        ],
        "title": "TransferRequest",
        "description": "转存请求"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}